+++ CHANGELOG.md
@@ -2,11 +2,13 @@
 
 ## NEXT
 ### ✨ Features and improvements
-- None
+- Adds `enableHaloAnimations`, `enableSpaceAnimations`, `disableHaloAnimations`, `disableSpaceAnimations` to Map instance for toggling halo and space animations.
 
 ### 🐛 Bug fixes
 - fixes a bug on `halo` where sharp edged stops cause noise to surface when scaled down.
 - fixes bug where space would not load correctly when style is a URI.
+- fixes a bug in the UMD module where space crashes when only a color is set.
+- fixes a bug when space: true and no style is passed, space becomes unset.
 
 ### Others
 - None